hw: add local core config parameter
Add local core ID to specify the core targeted by remote.
This is used only on S32xx platforms where Linux may be
running SMP and the pending interrupt needs to be cleared
in the targeted core instead of the executing core.

In ipc-uio also added remote core type to be able to
select the remote core index.

Issue: IPCF-1017
Signed-off-by: Manuel Arguelles <manuel.arguelles@nxp.com>