shm: add local trusted cores config
Add trusted core mask to local_core config to select
the cores that should be marked as trusted in the MSCM
IRCPCFG register. This is useful for example, when each
A53 cluster is running a Linux VM and not all cores
should be trusted.

Processor index is now a enum that can be used to
create the trusted core mask.

Issue: IPCF-1017
Signed-off-by: Manuel Arguelles <manuel.arguelles@nxp.com>